Restructured for other pages and currently working on teacher page
