+++ src/App.jsx
@@ -1,9 +1,15 @@
-import LandingPage from './components/LandingPage';
+import LandingPage from './components/LandingPage/LandingPage';
+
+//Mock Components
+//Teacher
+import TeacherAdminPage from './components/TeacherAdminPage/TeacherAdminPage';
+//Student
 
 function App() {
   return (
     <>
       <LandingPage />
+      <TeacherAdminPage />
     </>
   );
 }
