+++ src/components/TeacherAdminPage/TeacherAdminPage.jsx
@@ -1,0 +1,11 @@
+import Header from '../Header';
+
+const TeacherAdminPage = () => {
+  return (
+    <>
+      <Header />
+    </>
+  );
+};
+
+export default TeacherAdminPage;
